+++ pkg/cache/cache.go
@@ -6,4 +6,6 @@ type Cache interface {
 	RemoveReport(id string)
 	AddReport(report v1alpha2.ReportInterface)
 	GetResults(id string) []string
+	Shared() bool
+	Clear()
 }