+++ pkg/cache/redis.go
@@ -2,11 +2,12 @@ package cache
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
+	"strings"
 	"time"
 
 	goredis "github.com/go-redis/redis/v8"
+	"go.uber.org/zap"
 
 	"github.com/kyverno/policy-reporter/pkg/crd/api/policyreport/v1alpha2"
 )
@@ -18,28 +19,73 @@ type redisCache struct {
 }
 
 func (r *redisCache) AddReport(report v1alpha2.ReportInterface) {
-	list := reportResultsIds(report)
+	next := make(map[string]bool)
 
-	value, _ := json.Marshal(list)
+	for _, result := range report.GetResults() {
+		r.rdb.Set(context.Background(), r.generateKey(report.GetID(), result.GetID()), nil, 0)
+		next[result.GetID()] = true
+	}
 
-	r.rdb.Set(context.Background(), r.generateKey(report.GetID()), string(value), 0)
+	for _, id := range r.GetResults(report.GetID()) {
+		if !next[id] {
+			r.rdb.Set(context.Background(), r.generateKey(report.GetID(), id), nil, 6*time.Hour)
+		}
+	}
 }
 
 func (r *redisCache) RemoveReport(id string) {
-	r.rdb.Del(context.Background(), r.generateKey(id))
+	keys, err := r.rdb.Keys(context.Background(), r.generateKeyPattern(id)).Result()
+	if err != nil {
+		zap.L().Error("failed to load report keys", zap.Error(err))
+		return
+	}
+
+	for _, key := range keys {
+		r.rdb.Expire(context.Background(), key, 10*time.Minute)
+	}
 }
 
 func (r *redisCache) GetResults(id string) []string {
-	list, _ := r.rdb.Get(context.Background(), r.generateKey(id)).Result()
 	results := make([]string, 0)
+	pattern := r.generateKeyPattern(id)
 
-	json.Unmarshal([]byte(list), &results)
+	keys, err := r.rdb.Keys(context.Background(), pattern).Result()
+	if err != nil {
+		zap.L().Error("failed to load report keys", zap.Error(err))
+		return results
+	}
+
+	prefix := strings.TrimSuffix(pattern, "*")
+	for _, key := range keys {
+		results = append(results, strings.TrimPrefix(key, prefix))
+	}
 
 	return results
 }
 
-func (r *redisCache) generateKey(id string) string {
-	return fmt.Sprintf("%s:%s", r.prefix, id)
+func (r *redisCache) Shared() bool {
+	return true
+}
+
+func (r *redisCache) Clear() {
+	results := r.rdb.Keys(context.Background(), r.prefix+":*")
+
+	keys, err := results.Result()
+	if err != nil {
+		zap.L().Error("failed to find cache keys in redis", zap.Error(err))
+	}
+
+	for _, key := range keys {
+		r.rdb.Del(context.Background(), key)
+	}
+}
+
+func (r *redisCache) generateKey(report, id string) string {
+	return fmt.Sprintf("%s:%s:%s", r.prefix, report, id)
+}
+
+func (r *redisCache) generateKeyPattern(report string) string {
+	return fmt.Sprintf("%s:%s:*", r.prefix, report)
 }
 
 func NewRedisCache(prefix string, rdb *goredis.Client, ttl time.Duration) Cache {