+++ pkg/listener/store.go
@@ -1,6 +1,8 @@
 package listener
 
 import (
+	"context"
+
 	"go.uber.org/zap"
 
 	"github.com/kyverno/policy-reporter/pkg/report"
@@ -8,19 +10,19 @@ import (
 
 const Store = "store_listener"
 
-func NewStoreListener(store report.PolicyReportStore) report.PolicyReportListener {
+func NewStoreListener(ctx context.Context, store report.PolicyReportStore) report.PolicyReportListener {
 	return func(event report.LifecycleEvent) {
 		if event.Type == report.Deleted {
-			logOnError("remove", event.PolicyReport.GetName(), store.Remove(event.PolicyReport.GetID()))
+			logOnError("remove", event.PolicyReport.GetName(), store.Remove(ctx, event.PolicyReport.GetID()))
 			return
 		}
 
 		if event.Type == report.Updated {
-			logOnError("update", event.PolicyReport.GetName(), store.Update(event.PolicyReport))
+			logOnError("update", event.PolicyReport.GetName(), store.Update(ctx, event.PolicyReport))
 			return
 		}
 
-		logOnError("add", event.PolicyReport.GetName(), store.Add(event.PolicyReport))
+		logOnError("add", event.PolicyReport.GetName(), store.Update(ctx, event.PolicyReport))
 	}
 }