+++ pkg/report/client.go
@@ -1,6 +1,8 @@
 package report
 
-import "github.com/kyverno/policy-reporter/pkg/crd/api/policyreport/v1alpha2"
+import (
+	"github.com/kyverno/policy-reporter/pkg/crd/api/policyreport/v1alpha2"
+)
 
 // PolicyReportListener is called whenever a new PolicyReport comes in
 type PolicyReportListener = func(LifecycleEvent)
@@ -16,4 +18,6 @@ type PolicyReportClient interface {
 	Sync(stopper chan struct{}) error
 	// HasSynced the configured PolicyReport
 	HasSynced() bool
+	// Stop the client
+	Stop()
 }